Add a vorbiscomment-style metadata header to the theora bitstream format. As in vorbis, it's the second independent but required header packet, between the ident and info header and the decoder tables.
The new file comment.c has functions for manipulating the new
theora_comment structure. These and the comment structure are copied
directly from the vorbis source; only the names have been changed. Since
many xiph codecs share this format it's worth thinking about moving to
libogg (or liboggfile). There remains however a need for codec-specific
wrappers because all the codecs use a slightly different preamble.

There are several decisions made here. Like in vorbis, the
theora_comment structure is not associated with the info or codec state.
This makes for more housekeeping on the player side, but is consistent.
The justification in the vorbis source is 'so vorbis_info can be
declared static'. Sure.

Not so nice is that the separate parsing functions for each header
packet have been retained. This was reasonable for one and two header
packets, but I do think it makes too much work for the client with
three and we should move to a single theora_header_in() like vorbis
uses. This would take care of the parsing in order: the client just
calls it three times. Note that this is simpler if, contrary to the
above, the theora_comment is tracked inside the info or codec state
structures.

Finally, there's a question of endianness. Vorbis, the original ogg,
and speex are all little endian; theora is big endian. Should the
theora comment header by little or big endian? I picked little for
consistency with the other codecs, but it makes for a point of confusion
in the theora spec.


git-svn-id: http://svn.xiph.org/trunk/theora@4694 0101bb08-14d6-0310-b084-bc0e0c8e3800