Try to find installed versions of our .m4 macro files.
This is an experiment. We've been just manually copying ogg.m4,
vorbis.m4, etc. to the acinclude.m4 of whatever package needed them.
This is dorky, because the makefiles actually install them in
$(datadir)/aclocal, and the whole point of aclocal is to generate
aclocal.m4 with all the required macros from their installed
definitions.

Unfortunately, on my debian system at least, aclocal doesn't look in
/usr/local/share/aclocal by default. This patch attempts to find them in
some common locations and adds that directory to the aclocal invocation
in autogen.sh.

This is a complete hack, because of course we don't know $prefix at
autogen time. It makes it harder for developers using cvs, since things
can't be built if the macros aren't found, but release users are
unaffected because distributions will include them in the generated
aclocal.m4.


git-svn-id: http://svn.xiph.org/trunk/theora@4915 0101bb08-14d6-0310-b084-bc0e0c8e3800