Use the proper header path to SDL.h and use SDL_CFLAGS to add it to the search path.
The helper sdl-config script assumes you will use #include <SDL.h> and
puts an include path including the SDL/ directory into the CFLAGS. On
linux we were ignorning the output of sdl-config and including
<SDL/SDL.h> on the assumption that would work from one of the default
include paths (/usr/include or /usr/local/include). This is needed to
support platforms like macosx where this isn't generally true and also
qualifies as a code simplification since we were doing the right thing
on windows anyway.

Unfortunately the way $(SDL_CFLAGS) gets added to the compiler line is a
complete hack. This patch adds it to the global CFLAGS for everything in
the examples directory, when of course encoder_example and dump_video do
not require it. I shouldn't do any harm, but is conceptually wrong.

Current versions of automake support target-specific _CFLAGS lines, but
that doesn't work in automake 1.4, so it's either this or require
at least version 1.6.


git-svn-id: http://svn.xiph.org/trunk/theora@4934 0101bb08-14d6-0310-b084-bc0e0c8e3800