avi2vp3 cleanup.
open() on win32 requires an O_BINARY flag. If this flag isn't defined by
the os, we define it to 0 so that it has no effect when or'd with the
other flags.

win32 also doesn't define to the mode flags, so fall back to a manual
0644 constant.

AVI_max_video_chunk() returns zero on files without an idx table: fall back
to a 128K buffer in such cases. This is cleaner than the universal 32K buffer
used in the theora-tools version.

Based on a patch from Nehal Mistry.


git-svn-id: http://svn.xiph.org/trunk/theora@8991 0101bb08-14d6-0310-b084-bc0e0c8e3800