Rename the 'ystride' member of the th_img_plane struct to 'stride'.
The y was for vertical, to distinguish from a possible future xstride
to allow pixel packing, but this was easily confused with the y of
Y'CrCb color planes. This is less confusing, and I think it makes
more sense to add support for chunked images in a deparate type.

THIS IS AN INCOMPATIBLE CHANGE in the new theora-exp API.


git-svn-id: http://svn.xiph.org/trunk/theora@14385 0101bb08-14d6-0310-b084-bc0e0c8e3800