Bump the version number and change log for the 1.1.0 release.
Note that we're switching to a three-part version number here, like the 
other xiph code projects use. This has some implications for pkg-config. 
There was a problem with the 1.0 release because pkg-config sorted '1.0' 
later than '1.0beta3'. To work around this, the pkg-config files for the
1.0 release give the version at '1.0final'. Anything with 1.1 version 
sorts after 1.0, but 1.1.x doesn't sort after the 1.1beta versions we've 
already released. The 1.1 beta period was much shorter than for 1.0, so 
I've switched schemes for this release. Requesting 'theoradec >= 
1.0final' or 'theoradec >= 1.1.0' will work, but not, for 
example, 'theoradec >= 1.1beta2'.


git-svn-id: http://svn.xiph.org/trunk/theora@16561 0101bb08-14d6-0310-b084-bc0e0c8e3800