Require automake 1.7 which is the first version to support docdir.
Use of the built-in --docdir option was assumed by r16634, but systems 
with automake-1.6 installed would warn that it was unsupported. Updating 
the minimum version to 1.7 ensures a supporting automake is called.

Note this means we require at least MacOS X 10.5 to build out of 
the box on that platform.


git-svn-id: http://svn.xiph.org/trunk/theora@16670 0101bb08-14d6-0310-b084-bc0e0c8e3800