First drop of Theorarm code, tidied up a bit to be more suitable for merging. This code is still under GNU GPL, not BSD, but this is anticipated to change within a month or so.
This merged version of the code is currently untested, as I haven't got
cross compiling working with the standard libTheora makefiles yet, but is
(I hope) a reasonable first step.

The ARM code is dropped in pretty much unchanged from Theorarm, the C has
been tweaked a bit. The bitstream handling code from Theorarm is not in this
code yet, as libTheora is written to work with the current libOgg, and my
bitreading code is written to work with the version of libOgg present in
Tremor/Tremolo.

The C changes here make sense for ARM, but may not be ideal for other
architectures - additional work will be required to resolve this. It
may be that we need to discard some of these changes, or fork routines
for different architectures.



git-svn-id: http://svn.xiph.org/branches/theorarm-merge-branch@16954 0101bb08-14d6-0310-b084-bc0e0c8e3800