[Reapply of revision 17171 that went in without a commit message, and was therefore subsequently backed out in 17172.]
Tweak theorarm-merge-branch so that it still builds on linux using
./autogen.sh. This has meant pushing various of the oc_mv changes through
the encoder side too - this needs to be discussed with the library
maintainers as I fear they may not like it.

I can't see any other way to do it without losing speed on ARMs, but maybe
smarter people than me can spot something.

Next step is to put the bitreading changes in protected with OC_LIBOGG2.



git-svn-id: http://svn.xiph.org/branches/theorarm-merge-branch@17173 0101bb08-14d6-0310-b084-bc0e0c8e3800