Initial port of Robin Watts's Theorarm optimizations.
This includes the assembly for motion compensation, the iDCT, and the loop
 filter, with minor fixes, additions, and speed improvements.
David Schleef and Ralph Giles contributed to the autoconf and automake support.
Cristian Adam contributed to testing the CPU detection on Windows.

New assembly was written for the libtheora bitpacker, as well as new ARMv6
 versions of the iDCT and loop filter, and a new NEON version of the iDCT.
The iDCT was also modified to zero out its input coefficients afterwards, so
 they can be used for the next block, avoiding the need to clear many
 coefficients which were already zero (as an alternative to Robin's
 oc_memzero16_64, which always cleared all the coefficients).
CPU detection for Linux and Windows was written to allow execution of a single
 binary on multiple ARM CPUs; iOS detection is needed.
We currently do not try to detect the _minimum_ CPU the code is being compiled
 for, and thus will include some code for processors older than is strictly
 needed, though it is certainly possible to fix this.

A number of other updates were applied to the optimizations for other
 architectures to adapt to interface changes that helped ARM, and some
 long-standing bugs were fixed (such as an off-by-one in the iDCT for every
 architecture that was harmless, but hurt performance).
The x86 Windows assembly had bitrotted a bit, and has been restored closer to
 something that should work, but is completely untested.

What has _not_ been ported over yet is assembly for some of the less-frequently
 used zero-filling, border padding, coded block flag reading, the libogg2
 bitpacker, any of the post-processing, as well as most of the C code changes,
 in particular the MV handling and the use of pointer arithmetic instead of
 indexing in loops.
Some or all of this may follow in subsequent commits.
The current code, however, already benchmarks faster than the original Theorarm
 version on ARMv6 or NEON-capable hardware.
Some of that is due to the new routines written for these CPUs, while the rest
 is due to other general decoder improvemnts since the 1.1 release Theorarm was
 based on.


git-svn-id: http://svn.xiph.org/experimental/derf/theora-ptalarbvorm@17410 0101bb08-14d6-0310-b084-bc0e0c8e3800