Rename OC_SAD_SHIFT to OC_SATD_SHIFT.
The SAD and SADT metrics have different ranges, so it is
necessary to use different scale factors when comparing
them.

Originally we used SAD, and when SATD was added, the
scale factor was updated without changing the name.
This commit finally does so. However, it is intended
that the code switch back to the SAD metric at higher
speed levels, we also keep OC_SAD_SHIFT with its
appropriate value, 6 according to Tim.

Likewise, OC_SAD_BINS is renamed OC_COMP_BINS in
expectation of supporting mode tables for both
metrics.

git-svn-id: http://svn.xiph.org/trunk/theora@17645 0101bb08-14d6-0310-b084-bc0e0c8e3800